Use reactstrap, drop jsquery dep, add a "adaptative" table css
Reactstrap allows better control of the dropdowns toggling, which
is necessary for the "click to confirm action" feature added by
this patch.

Reactstrap also permits dropping the jquery dependency, which is
quite heavy.

Dangerous actions now require 6 confirmation clicks, and
impacting action 3.

This patch also factorize the ActionsDropdown is a common react
component.

The adaptative tables switch to the flex renderer when width
constraint would place part of the table outside the screen area.